fix initialization routines
in initialization, we need to update all of the leaf inputs. sometimes,
there are leaf inputs with overlapping outputs. previously, we would
trigger an update for each input, which would then trigger an update
for each of its outputs. now, we skip triggering updates to outputs if
they have already been covered by a separate input.